requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo[srv]==4.6.1
werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
pytest>=8.0.0
pytest-playwright>=0.7.0
